refactor(cpuhealth): split the Linux sampler into a cgroup source and a host source
The sampler read both a cgroup's own accounting files and machine-wide
host files under one type, but its doc said it only read a cgroup. Split
the reads into two unexported sources — cgroupSource (cpu.max, cpu.stat,
cpu.pressure, cpuset.cpus.effective) and hostSource (/proc/stat,
/proc/cpuinfo, the DMI identity files) — each owning the sticky state and
baseline derived from its own files, with linuxSampler now a thin
composer that stamps the tick's single Timestamp and derives CPU scope
from both sources' reads. Fixed the Sampler interface doc to describe
what it actually reads.

NewLinuxSampler and the Sampler interface keep their exact signatures;
this is a behaviour-preserving split, not a functional change.
