+++ umh-core/pkg/cpuhealth/host_source.go
@@ -12,15 +12,137 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Deciding whether this machine is a virtual machine, which is the fact that
-// makes stolen CPU time worth measuring at all.
+// The host source: everything this package reads machine-wide, independent
+// of any particular cgroup — /proc/stat, /proc/cpuinfo, and the DMI identity
+// files. Distinct from cgroupSource, which reads one cgroup's own accounting
+// files under its base.
 
 package cpuhealth
 
 import (
 	"context"
+	"strconv"
 	"strings"
+	"time"
+
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/diagnosis"
+	"github.com/united-manufacturing-hub/united-manufacturing-hub/umh-core/pkg/service/filesystem"
 )
+
+// userHz matches the kernel's USER_HZ: the tick rate dividing /proc/stat jiffies
+// into seconds, hence cores. It is hardcoded rather than queried because the
+// Linux userspace ABI fixes USER_HZ at 100 regardless of the kernel's internal
+// CONFIG_HZ, and the shipped binary builds with CGO_ENABLED=0 (see the
+// Dockerfile), so sysconf(_SC_CLK_TCK) is not reachable to ask instead.
+const userHz = 100.0
+
+// hostSource reads the machine-wide files: /proc/stat, /proc/cpuinfo,
+// /sys/class/dmi/id/product_name and /sys/class/dmi/id/sys_vendor. It owns
+// the two facts that persist across ticks for this host — the
+// host-busy/steal baseline and the sticky virtualisation fact — so it is
+// constructible and testable independently of cgroupSource.
+type hostSource struct {
+	fs filesystem.Service
+
+	hostBase hostBaseline
+
+	// virtualized is the sticky virtualisation fact, resolved on the first
+	// successful /proc/cpuinfo read and re-published without re-reading. An
+	// unreadable cpuinfo resolves false but is not cached, so a later readable
+	// one is still considered.
+	virtualized  bool
+	virtResolved bool
+}
+
+// newHostSource returns a hostSource reading via fs.
+func newHostSource(fs filesystem.Service) *hostSource {
+	return &hostSource{fs: fs}
+}
+
+// hostBaseline is the previous tick's /proc/stat edges (busy, steal, denominator
+// jiffy totals) and its read timestamp, from which hostRates derives the
+// instantaneous host-busy rate and per-interval steal fraction. have is false
+// until a first successful read fixes the baseline; a falling counter (a host
+// restart) re-baselines instead of publishing a nonsense value.
+type hostBaseline struct {
+	busy, steal, denom float64
+	time               time.Time
+	have               bool
+}
+
+// hostRates derives this tick's HostBusy rate and Steal fraction from busy,
+// steal and denom against the baseline this source owns, then updates the
+// baseline for the next tick. ts is the composer's single per-tick Timestamp,
+// passed in rather than read via time.Now(): this rate and the cgroup
+// source's usage rate must be measured against the same instant, or Decide's
+// attribution would compare a machine-wide mean against a cgroup mean taken
+// at a different moment.
+func (h *hostSource) hostRates(ts time.Time, busy, steal, denom float64) (hostBusy, stealFrac diagnosis.Reading) {
+	hostBusy = diagnosis.Unknown()
+	stealFrac = diagnosis.Unknown()
+	if h.hostBase.have {
+		// HostBusy: busy-jiffy delta ÷ USER_HZ ÷ elapsed seconds; skipped on
+		// a counter reset or zero elapsed time.
+		if busy >= h.hostBase.busy {
+			if elapsed := ts.Sub(h.hostBase.time).Seconds(); elapsed > 0 {
+				hostBusy = diagnosis.Known((busy - h.hostBase.busy) / userHz / elapsed)
+			}
+		}
+		// The steal fraction is the interval's steal-jiffy delta over the
+		// interval's total-jiffy delta. A falling steal counter (a reset)
+		// and a non-positive denominator delta (proc/stat did not advance,
+		// or is all zeros) publish nothing, never a NaN/Inf reading.
+		dDenom := denom - h.hostBase.denom
+		if steal >= h.hostBase.steal && dDenom > 0 {
+			stealFrac = diagnosis.Known((steal - h.hostBase.steal) / dDenom)
+		}
+	}
+	h.hostBase = hostBaseline{busy: busy, steal: steal, denom: denom, time: ts, have: true}
+	return hostBusy, stealFrac
+}
+
+// readHost reads the first aggregate "cpu " line of /proc/stat and yields the
+// raw busy, steal and denominator jiffy totals. busy counts user, nice, system,
+// irq and softirq jiffies (idle, iowait, steal, guest and guest_nice excluded).
+// The steal denominator is the sum of fields 0..7 only, since the kernel folds
+// guest and guest_nice into user and nice. Both totals are kept raw so the
+// caller can derive interval deltas. machine is the number of per-CPU (cpu0,
+// cpu1, …) lines in the same file, the machine's CPU count. The trailing space
+// in "cpu " keeps the aggregate line from matching cpu0/cpu1.
+func (h *hostSource) readHost(ctx context.Context) (busy, steal, denom, machine float64, ok bool) {
+	data, err := h.fs.ReadFile(ctx, "/proc/stat")
+	if err != nil {
+		return 0, 0, 0, 0, false
+	}
+	for _, line := range strings.Split(string(data), "\n") {
+		// A per-CPU line is "cpu" followed by a digit; the aggregate "cpu " line
+		// (space, not digit) is not one of them.
+		if len(line) > 3 && strings.HasPrefix(line, "cpu") && line[3] >= '0' && line[3] <= '9' {
+			machine++
+		}
+	}
+	for _, line := range strings.Split(string(data), "\n") {
+		if !strings.HasPrefix(line, "cpu ") {
+			continue
+		}
+		fields := strings.Fields(line) // fields[0] == "cpu"
+		if len(fields) < 9 {
+			return 0, 0, 0, machine, false
+		}
+		vals := make([]float64, len(fields))
+		for i := 1; i < len(fields); i++ {
+			v, err := strconv.ParseFloat(fields[i], 64)
+			if err != nil {
+				return 0, 0, 0, machine, false
+			}
+			vals[i] = v
+		}
+		busy := vals[1] + vals[2] + vals[3] + vals[6] + vals[7]
+		denom := vals[1] + vals[2] + vals[3] + vals[4] + vals[5] + vals[6] + vals[7] + vals[8]
+		return busy, vals[8], denom, machine, true
+	}
+	return 0, 0, 0, machine, false
+}
 
 // readVirtualized returns the sticky virtualisation fact. A "hypervisor" flag
 // in /proc/cpuinfo's flags line proves an x86 guest and caches true; a positive
@@ -32,14 +154,14 @@ import (
 // is a conclusive bare-metal identity. It stays open for the next tick —
 // never a permanent Virtualized=false — when no DMI source was readable, or on
 // a platform without a flags line when sys_vendor is still unresolved.
-func (s *linuxSampler) readVirtualized(ctx context.Context) bool {
-	if s.virtResolved {
-		return s.virtualized
+func (h *hostSource) readVirtualized(ctx context.Context) bool {
+	if h.virtResolved {
+		return h.virtualized
 	}
-	data, err := s.fs.ReadFile(ctx, "/proc/cpuinfo")
+	data, err := h.fs.ReadFile(ctx, "/proc/cpuinfo")
 	if err == nil && cpuinfoHasHypervisorFlag(data) {
-		s.virtualized = true
-		s.virtResolved = true
+		h.virtualized = true
+		h.virtResolved = true
 		return true
 	}
 	// ARM64 route. A successful DMI read resolves the fact either way; a failed
@@ -47,11 +169,11 @@ func (s *linuxSampler) readVirtualized(ctx context.Context) bool {
 	// spans two independent sources: product_name and sys_vendor (the cloud
 	// hypervisor an ARM64 product_name like "m6g.medium" never names), each read
 	// separately so a failure or absence on one never breaks the other.
-	pv, pok := s.dmiProductVirtualized(ctx)
-	vv, vok := s.dmiVendorVirtualized(ctx)
+	pv, pok := h.dmiProductVirtualized(ctx)
+	vv, vok := h.dmiVendorVirtualized(ctx)
 	if (pok && pv) || (vok && vv) {
-		s.virtualized = true
-		s.virtResolved = true
+		h.virtualized = true
+		h.virtResolved = true
 		return true
 	}
 	// Neither DMI source was readable — there is no evidence of a guest or of a
@@ -72,8 +194,8 @@ func (s *linuxSampler) readVirtualized(ctx context.Context) bool {
 	if (err != nil || !cpuinfoHasFlagsLine(data)) && !vok {
 		return false
 	}
-	s.virtualized = false
-	s.virtResolved = true
+	h.virtualized = false
+	h.virtResolved = true
 	return false
 }
 
@@ -129,8 +251,8 @@ var dmiHypervisorTokens = []string{
 // false when the read failed: an unreadable product_name is no evidence, and
 // leaving it unresolved lets the caller (and a later tick) retry rather than
 // caching Virtualized=false forever.
-func (s *linuxSampler) dmiProductVirtualized(ctx context.Context) (virtualized, resolved bool) {
-	data, err := s.fs.ReadFile(ctx, "/sys/class/dmi/id/product_name")
+func (h *hostSource) dmiProductVirtualized(ctx context.Context) (virtualized, resolved bool) {
+	data, err := h.fs.ReadFile(ctx, "/sys/class/dmi/id/product_name")
 	return dmiTokenMatch(data, err, dmiHypervisorTokens)
 }
 
@@ -152,8 +274,8 @@ var dmiVendorHypervisorTokens = []string{
 // dmiVendorVirtualized is the second ARM64 DMI source. It reports whether
 // /sys/class/dmi/id/sys_vendor names a cloud hypervisor, with the same
 // read-failure contract as dmiProductVirtualized.
-func (s *linuxSampler) dmiVendorVirtualized(ctx context.Context) (virtualized, resolved bool) {
-	data, err := s.fs.ReadFile(ctx, "/sys/class/dmi/id/sys_vendor")
+func (h *hostSource) dmiVendorVirtualized(ctx context.Context) (virtualized, resolved bool) {
+	data, err := h.fs.ReadFile(ctx, "/sys/class/dmi/id/sys_vendor")
 	return dmiTokenMatch(data, err, dmiVendorHypervisorTokens)
 }
