--- umh-core/pkg/cpuhealth/read_psi.go
@@ -1,51 +1,0 @@
-// Copyright 2025 UMH Systems GmbH
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Pressure Stall Information is the kernel's own measure of how long runnable
-// tasks waited for a core; this reads the cgroup's figure. The interface is
-// documented at https://www.kernel.org/doc/html/latest/accounting/psi.html.
-
-package cpuhealth
-
-import (
-	"context"
-	"strconv"
-	"strings"
-)
-
-// readPSI reads cpu.pressure's "some" avg60 as a 0..1 fraction.
-func (s *linuxSampler) readPSI(ctx context.Context) (frac float64, ok bool) {
-	data, err := s.fs.ReadFile(ctx, s.base+"/cpu.pressure")
-	if err != nil {
-		return 0, false
-	}
-
-	for _, line := range strings.Split(string(data), "\n") {
-		if !strings.HasPrefix(line, "some") {
-			continue
-		}
-		for _, field := range strings.Fields(line) {
-			if strings.HasPrefix(field, "avg60=") {
-				v, err := strconv.ParseFloat(strings.TrimPrefix(field, "avg60="), 64)
-				if err != nil {
-					// An unparsable avg60 is no pressure this tick, matching the
-					// unparsable cpu.max no-signal handling: never a present 0.0.
-					return 0, false
-				}
-				return v / 100.0, true
-			}
-		}
-	}
-	return 0, false
-}
